+++ docs/setup/changelog.rst
@@ -21,6 +21,7 @@ Upcoming Changes (main branch)
 - Added ``array.copy()`` and ``nd.copy(array)`` functions.
 - Added ``positive`` and ``negative`` functions.
 - Added ``count_nonzero`` functions.
+- Added ``concatenate`` functions.
 
 **Fixed**